requirements.txt
Django~=4.2.4
Pillow~=10.0.0
django-admin-sortable2~=2.1.8
django-tinymce~=3.6.1
django-environ~=0.10.0